amoeba/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "amoeba"
version = "0.1.0"
description = "Amoeba is an agent framework."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "litellm>=1.82.6",
    "python-dotenv>=1.2.2",
]

# Real package lives under src/amoeba/ (do not add __init__.py at repo root).
[tool.setuptools.packages.find]
where = ["src"]
